+++ test/astropy.jl
@@ -3,6 +3,8 @@
 using LombScargle, Base.Test, PyCall
 PyCall.@pyimport astropy.stats as ast
 
+srand(1)
+
 ntimes = 401
 t = linspace(0.01, 10pi, ntimes)
 t += step(t)*rand(ntimes)
@@ -24,25 +26,47 @@ for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3)
     end
 end
 
-# Test the fast method with evenly spaced data.
+# Test the fast method with evenly spaced data.  Use both heteroskedastic and
+# homoskedastic uncertainties.
 t = linspace(0.01, 10pi, ntimes)
-for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3)
+errors = rand(0.1:1e-3:4.0, ntimes)
+for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3), err in (ones(ntimes), errors)
     s = f(t)
     # "psd" normalization in LombScargle slightly differ from that of
     # Astropy and the test would fail if we includ it.
     for fitmean in (true, false), nrm in ("standard", "model", "log")
-        f_jl, p_jl = freqpower(lombscargle(t, s, fit_mean = fitmean,
+        f_jl, p_jl = freqpower(lombscargle(t, s, err,
+                                           fit_mean = fitmean,
                                            normalization=nrm,
                                            maximum_frequency=10,
                                            samples_per_peak=10))
         f_py, p_py =
-            ast.LombScargle(t, s,
+            ast.LombScargle(t, s, dy = err,
                             fit_mean = fitmean)[:autopower](method="fast",
                                                             normalization=nrm,
                                                             maximum_frequency=10,
                                                             samples_per_peak=10)
         @test_approx_eq f_jl f_py
         @test_approx_eq p_jl p_py
+    end
+end
+
+# Test heteroskedastic uncertainties with non-fast method.  Test only standard
+# normalization, the only one for which LombScargle and Astropy give similar
+# results.
+for f in (x -> sinpi(x), x -> sin(x) + 1.5*cospi(4*x) + 3)
+    s = f(t)
+    for fitmean in (true, false)
+        f_jl, p_jl = freqpower(lombscargle(t, s, errors,
+                                           fit_mean = fitmean,
+                                           fast = false,
+                                           maximum_frequency=20))
+        f_py, p_py =
+            ast.LombScargle(t, s, dy = errors,
+                            fit_mean = fitmean)[:autopower](method="cython",
+                                                            maximum_frequency=20)
+        @test_approx_eq f_jl f_py
+        @test isapprox(p_jl, p_py, rtol = 0.08)
     end
 end